Refactored Navbar code for improved clarity and maintainability
